starter-kits/rag-production/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.0
qdrant-client==1.11.0
sentence-transformers==3.1.1
python-dotenv==1.0.1
pydantic==2.9.2
torch==2.4.1
tqdm==4.66.5
